JNG-2116 Add row property to all ui components JNG-2121 Add switch to components

+++ lib/src/judo_row.dart
@@ -14,7 +14,7 @@ class JudoRow extends StatelessWidget {
 
   final List<Widget> children;
   final int col;
-  final int row;
+  final double row;
   final MainAxisAlignment mainAxisAlignment;
   final CrossAxisAlignment crossAxisAlignment;
   final MainAxisSize mainAxisSize;
